# OTLP receiver: Optimize by initializing regexps at program start