Optimize the OTLP translation logic by initializing two regular expressions at program start, instead of during function calls. Thanks to @bboreham for [pointing out](https://github.com/prometheus/prometheus/pull/15710#discussion_r1899484120) this potential optimization.

#### Benchmark results

<details>
<summary>PrometheusConverter.FromMetrics benchmark stats</summary>

```console
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheusremotewrite
cpu: Intel(R) Xeon(R) Platinum 8280 CPU @ 2.70GHz
                                                                                                                                                               │ unoptimized.txt │           optimized.txt            │
                                                                                                                                                               │     sec/op      │   sec/op     vs base               │
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16            12.985m ± 4%   9.471m ± 2%  -27.06% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16             15.57m ± 5%   12.26m ± 3%  -21.24% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16            18.34m ± 3%   14.85m ± 4%  -19.00% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16            55.66m ± 2%   52.59m ± 2%   -5.52% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16            59.25m ± 3%   56.09m ± 1%   -5.32% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16           61.91m ± 2%   58.09m ± 2%   -6.16% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16             18.41m ± 3%   16.48m ± 7%  -10.46% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16             23.42m ± 5%   20.36m ± 2%  -13.06% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16            26.04m ± 4%   23.50m ± 7%   -9.75% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16            56.76m ± 4%   54.87m ± 3%   -3.32% (p=0.004 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16            61.43m ± 3%   59.57m ± 2%   -3.03% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16           64.34m ± 1%   62.58m ± 3%   -2.74% (p=0.009 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16          30.97m ± 2%   25.87m ± 6%  -16.46% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16          38.52m ± 3%   35.00m ± 4%   -9.15% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16         44.52m ± 3%   40.41m ± 3%   -9.23% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16         113.4m ± 1%   111.0m ± 2%   -2.06% (p=0.009 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16         120.2m ± 2%   116.7m ± 1%   -2.89% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16        126.7m ± 1%   124.2m ± 2%        ~ (p=0.065 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16                1.247µ ± 2%   1.276µ ± 1%        ~ (p=0.065 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16                1.264µ ± 1%   1.263µ ± 1%        ~ (p=0.563 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16               1.253µ ± 2%   1.258µ ± 2%        ~ (p=0.331 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16               1.245µ ± 2%   1.260µ ± 2%        ~ (p=0.329 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16               1.272µ ± 2%   1.242µ ± 2%   -2.32% (p=0.015 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16              1.261µ ± 1%   1.252µ ± 2%        ~ (p=0.058 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16            13.527m ± 3%   9.664m ± 2%  -28.56% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16             16.50m ± 3%   12.36m ± 5%  -25.07% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16            19.11m ± 2%   14.58m ± 3%  -23.74% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16            56.82m ± 2%   52.33m ± 1%   -7.90% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16            59.97m ± 1%   55.31m ± 2%   -7.77% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16           62.16m ± 3%   57.81m ± 2%   -7.01% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16             19.03m ± 4%   16.25m ± 4%  -14.65% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16             22.20m ± 3%   20.69m ± 3%   -6.79% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16            26.27m ± 3%   24.38m ± 3%   -7.21% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16            58.28m ± 3%   54.98m ± 2%   -5.66% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16            61.31m ± 4%   59.74m ± 4%        ~ (p=0.093 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16           64.76m ± 2%   64.14m ± 2%        ~ (p=0.394 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16          31.64m ± 6%   27.24m ± 5%  -13.92% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16          38.96m ± 3%   34.64m ± 4%  -11.09% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16         44.31m ± 1%   39.44m ± 2%  -10.99% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16         113.6m ± 1%   108.4m ± 1%   -4.57% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16         120.8m ± 3%   117.3m ± 3%   -2.89% (p=0.041 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16        127.0m ± 2%   121.6m ± 3%   -4.22% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16               1.245µ ± 2%   1.255µ ± 1%        ~ (p=0.310 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16               1.256µ ± 1%   1.251µ ± 1%        ~ (p=0.372 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16              1.248µ ± 2%   1.245µ ± 1%        ~ (p=0.485 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16              1.252µ ± 1%   1.252µ ± 1%        ~ (p=0.485 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16              1.263µ ± 1%   1.245µ ± 1%   -1.39% (p=0.006 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16             1.263µ ± 1%   1.245µ ± 2%        ~ (p=0.084 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16            13.69m ± 1%   10.08m ± 3%  -26.39% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16            17.12m ± 2%   12.60m ± 4%  -26.39% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16           19.78m ± 3%   15.72m ± 2%  -20.55% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16           57.93m ± 1%   54.62m ± 1%   -5.71% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16           62.41m ± 1%   57.93m ± 1%   -7.17% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16          65.03m ± 1%   60.36m ± 1%   -7.18% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16            19.74m ± 2%   17.72m ± 2%  -10.23% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16            24.30m ± 3%   22.43m ± 4%   -7.73% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16           27.54m ± 2%   25.24m ± 3%   -8.34% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16           59.36m ± 3%   56.89m ± 1%   -4.16% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16           64.20m ± 3%   60.69m ± 3%   -5.46% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16          68.01m ± 1%   64.14m ± 2%   -5.69% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16         33.82m ± 1%   26.77m ± 2%  -20.83% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16         41.23m ± 3%   33.67m ± 4%  -18.34% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16        47.60m ± 2%   39.60m ± 2%  -16.81% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16        115.3m ± 2%   107.3m ± 4%   -6.95% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16        123.3m ± 2%   114.7m ± 2%   -7.02% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_50/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16       127.9m ± 2%   124.2m ± 2%   -2.89% (p=0.002 n=6)
geomean                                                                                                                                                              6.449m        5.874m        -8.91%

                                                                                                                                                               │ unoptimized.txt │             optimized.txt             │
                                                                                                                                                               │      B/op       │     B/op      vs base                 │
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16            3.999Mi ± 0%   2.254Mi ± 1%  -43.64% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16            6.071Mi ± 0%   4.372Mi ± 0%  -27.99% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16           8.185Mi ± 0%   6.487Mi ± 0%  -20.75% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16          10.674Mi ± 0%   8.973Mi ± 0%  -15.93% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16           12.77Mi ± 0%   11.09Mi ± 0%  -13.15% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_0/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16          14.87Mi ± 0%   13.18Mi ± 0%  -11.32% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16            8.944Mi ± 0%   8.052Mi ± 2%   -9.98% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_5-16            11.84Mi ± 0%   10.97Mi ± 0%   -7.33% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_10-16           14.78Mi ± 0%   13.95Mi ± 0%   -5.61% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_0-16           24.07Mi ± 0%   23.23Mi ± 0%   -3.48% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_5-16           26.97Mi ± 0%   26.08Mi ± 0%   -3.30% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_0/labels_per_metric:_20/exemplars_per_series:_10-16          29.91Mi ± 0%   29.05Mi ± 0%   -2.86% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_0-16         12.64Mi ± 0%   10.02Mi ± 1%  -20.71% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_5-16         17.57Mi ± 0%   15.05Mi ± 0%  -14.33% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_2/exemplars_per_series:_10-16        22.65Mi ± 0%   20.12Mi ± 0%  -11.20% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_0-16        34.37Mi ± 0%   31.85Mi ± 0%   -7.33% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_5-16        39.32Mi ± 0%   36.80Mi ± 0%   -6.40% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_0/histogram_count:_1000/non-histogram_count:_1000/labels_per_metric:_20/exemplars_per_series:_10-16       44.40Mi ± 0%   41.92Mi ± 0%   -5.59% (p=0.002 n=6)
PrometheusConverter_FromMetrics/resource_attribute_count:_5/histogram_count:_0/non-histogram_count:_0/labels_per_metric:_2/exemplars_per_series:_0-16                 96.00 ± 0%     96.00 ± 0%        ~ (p=1.000 n=6) ¹
…