@bboreham 2024-12-30
I'm a bit suspicious that the biggest effect is 27%.  Is this a realistic benchmark?